sysaudio: target a desired buffer size / latency instead of assuming maximum
On most platforms the maximum amount of frames we can write is reasonable and
provides rarely noticable latency, but on some platforms (e.g. Windows prior
to Mason's patch) this is not true. It's a good idea therefor to target a
desired buffer size / number of frames to write per update, which is also exactly
what we must decide on the WebAudio backend. For now this is hard-coded in sysaudio
but we will expose this as a config option soon.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>