coreaudio: always use valid sample rate during createRecorder
`err_no -10863` is encountered when e.g. the builtin microphone
of a mac is configured (e.g. through Audio MIDI Setup) to use a different
format frequency, such as 48,000Hz, instead of what the application is trying
to capture (such as 44,100Hz.)

I tried to locate an API for configuring the sample rate of a device, but
it seems this is not straightforward through CoreAudio. During device discovery
(`Context.refresh`) we locate the microphone, but only with whatever sample rate
is configured already:

```
device 'MacBook Pro Microphone' channels:0 sample_rate: 48000 mode: playback
device 'MacBook Pro Microphone' channels:1 sample_rate: 48000 mode: capture
```

It seems we discover the `kAudioDevicePropertyNominalSampleRate` _nominal_ sample
rate, which appears to be the 'current nominal sample rate', i.e. what is configured
in the OS currently (as that needs no conversion/resampling, hence it is nominal.)

In practice, I don't think selecting the sample rate of a microphone is very
useful - it could be in some cases, but they are probably more niche, and one
can just configure it through the OS instead if that is desired.

This change makes us use the sample rate that is already in use, which means `err_no -10863`
goes away and creating recorders works on macOS as expected now, even if other
applications (e.g. OBS Studio) alter the sample rate.

Helps hexops/mach#880

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>