coreaudio: request microphone permission during createRecorder
Prior to this change, we would not request access to the microphone
during a call to createRecorder. In some situations, this would mean
you could record audio still (e.g. when the parent process, say a terminal,
was sufficiently privileged) - but as a standalone app, it was problematic.

In specific, this surfaced as 'I can only record silence'.

When `createRecorder` is called, we now properly prompt the user for microphone
permission.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>